making the world safer for running into a container

+++ remote.sh
@@ -24,14 +24,21 @@ fi
 
 function sshclean {
     echo "sshclean $1"
-    ssh-keygen -q -f "~/.ssh/known_hosts" -R `dig +short $1`
-    ssh-keygen -q -f "~/.ssh/known_hosts" -R $1
+    # ssh-keygen -q -f "~/.ssh/known_hosts" -R `dig +short $1`
+    # ssh-keygen -q -f "~/.ssh/known_hosts" -R $1
 }
 
 for node in "$@"
 do
     echo "$node (${user}${node})"
+
+    if [ -f $wd/password ]; then
+      which sshpass || ( echo "please install sshpass" ; exit )
+      sshpass -p `cat $wd/password` ssh-copy-id ${user}${node}
+    fi
+
     sshclean $node || echo "sshclean?"
+    echo "scp $wd/* ${user}${node}:"
     scp $wd/* ${user}${node}:
     ssh -t ${user}${node} sudo bash -ve do_it
 done
